Master update #2

+++ Headers/DC_CommandQueue.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 //function declerations
-void addToQueue(uint16_t channel, uint16_t dutyCycle, uint16_t direction, uint16_t time);
+void addToQueue(uint16_t channel, uint16_t dutyCycle, uint16_t direction, uint32_t time);
 struct queue* extractFromQueue(uint16_t channel);
 
 //structure for the queue
@@ -11,7 +11,7 @@ struct queue
 {
   struct queue* next;   //pointer to next node in queue
   uint16_t dutyCycle;
-  uint16_t time;
+  uint32_t time;
   uint16_t direction;
   uint16_t channel;
 };
